+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Removed locking as it was already handled in underlying [hdmimatrix](https://github.com/marklynch/hdmimatrix) code
+- Centralised matrix operations to the coordinator so that only one hdmimatrix is initialised
 
 ## [0.0.4] - 2025-10-09
 ### Changed